src/main/kotlin/com/artemas/dgs/data/ShowsDataFetcher.kt
package com.artemas.dgs.data

import com.artemas.dgs.domain.Show
import com.artemas.dgs.service.ShowsService
import com.netflix.graphql.dgs.DgsComponent
import com.netflix.graphql.dgs.DgsQuery
import com.netflix.graphql.dgs.InputArgument

@DgsComponent
class ShowsDataFetcher(
    val showsService: ShowsService
) {

    @DgsQuery
    fun shows(@InputArgument titleFilter : String?): List<Show> {
        return if(titleFilter != null) {
            showsService.shows().filter { it.title.contains(titleFilter) }
        } else {
            showsService.shows()
        }
    }
}